services_1/data_ingestion/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.1